requirements.txt
agno>=1.0.0
groq>=0.9.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.18.0
fpdf2>=2.7.0